now able to divide between floats

+++ reflow_server/formula/utils/builtins/objects/Float.py
@@ -84,7 +84,7 @@ class Float(Object):
         object_representation = obj._representation_()
 
         if obj.type in [FLOAT_TYPE, INTEGER_TYPE]:
-            if (int(object_representation) == 0):
+            if (float(object_representation) == 0):
                 from reflow_server.formula.utils.builtins.objects.Error import Error
                 Error(self.settings)._initialize_('Error', 'Cannot divide by 0')
             else:
@@ -110,7 +110,7 @@ class Float(Object):
         object_representation = obj._representation_()
 
         if obj.type in [FLOAT_TYPE, INTEGER_TYPE]:
-            if (int(object_representation) == 0):
+            if (float(object_representation) == 0):
                 from reflow_server.formula.utils.builtins.objects.Error import Error
                 Error(self.settings)._initialize_('Error', 'Cannot divide by 0')
             else:
